requirements.txt
aiohttp==3.6.2
async-timeout==3.0.1
asyncio==3.4.3
attrs==19.3.0
cachelib==0.1.1
certifi==2019.6.16
chardet==3.0.4
click==7.1.2
Cython==0.29.12
discord.py==1.3.4
Django==2.2.1
Flask==1.1.2
Flask-Session==0.3.2
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
multidict==4.7.6
numpy==1.16.4
oauthlib==3.1.0
pytz==2019.2
requests==2.24.0
requests-oauthlib==0.5.0
sqlparse==0.3.0
urllib3==1.25.10
websockets==8.1
Werkzeug==1.0.1
yarl==1.4.2
